chore(opytimizer): Using integer timestamps instead of float.

+++ tests/opytimizer/core/test_agent.py
@@ -130,7 +130,7 @@ def test_agent_ub_setter():
 def test_agent_ts():
     new_agent = agent.Agent(1, 1, 0, 1)
 
-    assert type(new_agent.ts) == float
+    assert type(new_agent.ts) == int
 
 
 def test_agent_ts_setter():
@@ -139,9 +139,9 @@ def test_agent_ts_setter():
     try:
         new_agent.ts = np.array([0])
     except:
-        new_agent.ts = 0.0
+        new_agent.ts = 0
 
-    assert new_agent.ts == 0.0
+    assert new_agent.ts == 0
 
 
 def test_agent_clip_by_bound():
